requirements.txt
click>=7.0.0
pyyaml>=5.0.0
requests
xtermcolor
